# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Food Hub API
  version: 1.0.0
  description: API for Food Hub application which is food delivery service.
servers:
  - url: http://localhost:4000/api/v1
    description: Local Development Server
  - url: http://api.foodhub.com/api/v1
    description: Production Server
...
paths:
...
  /admin/users:
    get:
      summary: Get all users
      operationId: getUsers
      security:
        - BearerAuth: []
      tags:
        - Admin - Users
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/searchParam'
        - $ref: '#/components/parameters/filterParam'
        - $ref: '#/components/parameters/sortByParam'
        - $ref: '#/components/parameters/sortTypeParam'
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    example: 200
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Users retrieved successfully
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '500':
          $ref: '#/components/responses/500'
    post:
      summary: Create a new user
      operationId: createUser
      security:
        - BearerAuth: []
      tags:
        - Admin - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    example: 201
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: User created successfully
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '409':
          $ref: '#/components/responses/409'
        '500':
          $ref: '#/components/responses/500'
...
components:
  schemas:
    Timestamps:
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - createdAt
        - updatedAt
    User:
      allOf:
        - type: object
          properties:
            id:
              type: string
              format: uuid
              example: 64b64c4f2f9b9c001234567
            name:
              type: string
              example: John Doe
            email:
              type: string
              example: john.doe@mail.com
              format: email
            role:
              type: string
              example: CUSTOMER
              enum: [CUSTOMER, PROVIDER, ADMIN]
            status:
              type: string
              example: ACTIVE
              enum: [PENDING, ACTIVE, SUSPENDED]
          required:
            - id
            - name
            - email
            - role
            - status
        - $ref: '#/components/schemas/Timestamps'
    UserInput:
      type: object
      properties:
        name:
          type: string
          example: John Doe
        email:
          type: string
          example: john.doe@mail.com
          format: email
        password:
          type: string
          example: password123
          format: password
        role:
          type: string
          example: CUSTOMER
          enum: [CUSTOMER, PROVIDER, ADMIN]
        status:
          type: string
          example: ACTIVE
          enum: [PENDING, ACTIVE, SUSPENDED]
      required:
        - name
        - email
        - password
        - role
        - status
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          format: int32
          example: 2
        pageSize:
          type: integer
          format: int32
          example: 10
        totalPages:
          type: integer
          format: int32
          example: 5
        totalItems:
          type: integer
          format: int32
          example: 50
        next:
          type: string
          format: uri
          example: /users?page=3&pageSize=10
        previous:
          type: string
          format: uri
          example: /users?page=1&pageSize=10
      required:
        - page
        - pageSize
        - totalPages
        - totalItems
  responses:
...
    '400':
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 400
              success:
                type: boolean
                example: false
              error:
                type: string
                example: Bad Request
              message:
                type: string
                example: The request could not be understood or was missing required parameters.
    '401':
      description: Unauthorized access
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 401
              success:
                type: boolean
                example: false
              error:
                type: string
                example: Unauthorized access
              message:
                type: string
                example: You don't have the right permissions to access this resource.
    '403':
      description: Forbidden - You don't have permission
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 403
              success:
                type: boolean
                example: false
              error:
                type: string
                example: Forbidden
              message:
                type: string
                example: Forbidden - You don't have permission
...
    '409':
      description: Conflict - Resource already exists
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 409
              success:
                type: boolean
                example: false
              error:
                type: string
                example: Conflict
              message:
                type: string
                example: Conflict - Resource already exists
    '500':
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 500
              success:
                type: boolean
                example: false
              error:
                type: string
                example: Internal Server Error
              message:
                type: string
                example: We are sorry for the inconvenience. please try again later
  parameters:
    pageParam:
      in: query
      name: page
      description: current page number
      schema:
        type: integer
        format: int32
        example: 2
    limitParam:
      in: query
      name: pageSize
      description: Number of items per page
      schema:
        type: integer
        format: int32
        example: 10
    sortTypeParam:
      in: query
      name: sortType
      description: The type of sort (asc or dsc)
      schema:
        type: string
        enum: [asc, desc]
        example: desc
    sortByParam:
      in: query
      name: sortBy
      description: Sort key
      schema:
        type: string
        example: updatedAt
    searchParam:
      in: query
      name: search
      description: Search term to filter resources
      schema:
        type: string
        example: pizza
    filterParam:
      in: query
      name: filter
      description: Filter key
      schema:
        type: string
        example: category
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT